+++ classes/commoneer/assets/interface.php
@@ -10,7 +10,9 @@ interface Commoneer_Assets_Interface
 {
 
 
-	/** Que a new javascript file
+	/**
+	 * Que a new javascript file
+	 *
 	 * @static
 	 * @abstract
 	 * @param $names
@@ -18,7 +20,9 @@ interface Commoneer_Assets_Interface
 	 */
 	public static function use_script($names);
 
-	/** Que a new stylesheet (by default, LESS)
+	/**
+	 * Que a new stylesheet (by default, LESS)
+	 *
 	 * @static
 	 * @abstract
 	 * @param $names
@@ -26,16 +30,28 @@ interface Commoneer_Assets_Interface
 	 */
 	public static function use_style($names);
 
-	/** Que a new CSS stylesheet
+	/**
+	 * Que a new preset
+	 *
 	 * @static
 	 * @abstract
 	 * @param $names
 	 * @return void
 	 */
-	public static function use_css($names);
+	public static function preset($names);
+
+	/**
+	 * Get the singleton instance
+	 *
+	 * @static
+	 * @abstract
+	 * @return Commoneer_Assets
+	 */
+	public static function instance();
 
 	/**
 	 * Output all qued files as HTML markup (used in template head)
+	 *
 	 * @static
 	 * @abstract
 	 * @param bool|string $type Render all or only one type of includes